app/requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
PyJWT==2.10.1
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
